# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zivoo API",
    "description": "Parent and device APIs for the Zivoo learning toy. All routes are versioned under /v1. Errors use {error: {code, message, details}}.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/sessions": {
      "get": {
        "tags": [
          "parent"
        ],
        "summary": "List Sessions",
        "operationId": "list_sessions_v1_sessions_get",
        "parameters": [
          {
            "name": "child_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Child Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SessionSummaryOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/device/claim": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "Device Claim",
        "operationId": "device_claim_v1_device_claim_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceClaimIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/device/heartbeat": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "Heartbeat",
        "description": "Report status and receive the latest desired config plus pending commands.\n\nProduction transport should be a persistent TLS connection (MQTT/WebSocket);\nthis polling endpoint carries the same messages.",
        "operationId": "heartbeat_v1_device_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceSyncOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/device/commands/{command_id}/ack": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "Ack",
        "operationId": "ack_v1_device_commands__command_id__ack_post",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Command Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandAckIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/device/sessions": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "Start Session",
        "operationId": "start_session_v1_device_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToyReplyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/device/sessions/{client_session_id}/turns": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "Child Turn",
        "operationId": "child_turn_v1_device_sessions__client_session_id__turns_post",
        "parameters": [
          {
            "name": "client_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Client Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildUtteranceIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToyReplyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/device/sessions/{client_session_id}/end": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "End Session",
        "description": "Child walked away or interrupted. Idempotent.",
        "operationId": "end_session_v1_device_sessions__client_session_id__end_post",
        "parameters": [
          {
            "name": "client_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Client Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToyReplyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/device/telemetry/latency": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "Latency",
        "operationId": "latency_v1_device_telemetry_latency_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LatencyBatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ActivityStartIn": {
        "properties": {
          "client_session_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Client Session Id"
          },
          "activity_key": {
            "type": "string",
            "enum": [
              "phonics",
              "vocabulary",
              "spelling",
              "maths",
              "stories"
            ],
            "title": "Activity Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "client_session_id",
          "activity_key"
        ],
        "title": "ActivityStartIn"
      },
...
      "ChildUtteranceIn": {
        "properties": {
          "turn_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Turn Id"
          },
          "transcript": {
            "type": "string",
            "maxLength": 400,
            "title": "Transcript"
          },
          "stt_confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stt Confidence"
          },
          "audio_quality": {
            "type": "string",
            "enum": [
              "good",
              "noisy",
              "clipped",
              "silent"
            ],
            "title": "Audio Quality",
            "default": "good"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turn_id",
          "transcript"
        ],
        "title": "ChildUtteranceIn",
        "description": "What the toy heard. Text comes from on-cloud STT; confidence from that STT."
      },
...
      "CommandAckIn": {
        "properties": {
          "result": {
            "type": "string",
            "enum": [
              "ok",
              "failed"
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "result"
        ],
        "title": "CommandAckIn"
      },
...
      "CommandKind": {
        "type": "string",
        "enum": [
          "apply_config",
          "play_greeting",
          "enter_wifi_setup",
          "factory_reset",
          "check_firmware"
        ],
        "title": "CommandKind"
      },
...
      "DeviceClaimIn": {
        "properties": {
          "claim_token": {
            "type": "string",
            "maxLength": 128,
            "minLength": 10,
            "title": "Claim Token"
          },
          "firmware_version": {
            "type": "string",
            "maxLength": 32,
            "title": "Firmware Version"
          },
          "wifi_ssid": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Wifi Ssid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "claim_token",
          "firmware_version"
        ],
        "title": "DeviceClaimIn"
      },
      "DeviceClaimOut": {
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "credential": {
            "type": "string",
            "title": "Credential"
          },
          "desired_config_version": {
            "type": "integer",
            "title": "Desired Config Version"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "credential",
          "desired_config_version"
        ],
        "title": "DeviceClaimOut"
      },
      "DeviceCommandOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/CommandKind"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "config_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "payload",
          "config_version"
        ],
        "title": "DeviceCommandOut"
      },
...
      "DeviceSyncOut": {
        "properties": {
          "desired_config_version": {
            "type": "integer",
            "title": "Desired Config Version"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "commands": {
            "items": {
              "$ref": "#/components/schemas/DeviceCommandOut"
            },
            "type": "array",
            "title": "Commands"
          }
        },
        "type": "object",
        "required": [
          "desired_config_version",
          "config",
          "commands"
        ],
        "title": "DeviceSyncOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeartbeatIn": {
        "properties": {
          "firmware_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Firmware Version"
          },
          "wifi_ssid": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Wifi Ssid"
          },
          "applied_config_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Config Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "HeartbeatIn"
      },
      "LatencyBatchIn": {
        "properties": {
          "samples": {
            "items": {
              "$ref": "#/components/schemas/LatencyIn"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Samples"
          }
        },
        "type": "object",
        "required": [
          "samples"
        ],
        "title": "LatencyBatchIn"
      },
      "LatencyIn": {
        "properties": {
          "metric": {
            "type": "string",
            "enum": [
              "speech_end_to_first_audio",
              "stt",
              "assessment",
              "response_generation",
              "tts_first_audio",
              "reconnect",
              "network_failure"
            ],
            "title": "Metric"
          },
          "value_ms": {
            "type": "number",
            "maximum": 600000.0,
            "minimum": 0.0,
            "title": "Value Ms"
          },
          "conditions": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metric",
          "value_ms"
        ],
        "title": "LatencyIn"
      },
...
      "Outcome": {
        "type": "string",
        "enum": [
          "correct",
          "needs_practice",
          "uncertain"
        ],
        "title": "Outcome"
      },
      "Page_SessionSummaryOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SessionSummaryOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[SessionSummaryOut]"
      },
...
      "SessionSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "child_id": {
            "type": "string",
            "format": "uuid",
            "title": "Child Id"
          },
          "activity_key": {
            "type": "string",
            "title": "Activity Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "needs_practice": {
            "type": "integer",
            "title": "Needs Practice"
          },
          "uncertain": {
            "type": "integer",
            "title": "Uncertain"
          }
        },
        "type": "object",
        "required": [
          "id",
          "child_id",
          "activity_key",
          "status",
          "started_at",
          "ended_at",
          "summary",
          "correct",
          "needs_practice",
          "uncertain"
        ],
        "title": "SessionSummaryOut"
      },
...
      "ToyReplyOut": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "say": {
            "type": "string",
            "title": "Say"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Outcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "done": {
            "type": "boolean",
            "title": "Done",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "state",
          "say"
        ],
        "title": "ToyReplyOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}